Move timing helpers to backend utils
