+++ backend/preset_editor.py
@@ -7,9 +7,8 @@ from core import (
     DEFAULT_DB_PATH,
     DEFAULT_SETS_PER_EXERCISE,
     DEFAULT_REST_DURATION,
-    _to_db_timing,
-    _from_db_timing,
 )
+from backend.utils import _to_db_timing, _from_db_timing
 
 
 class PresetEditor:
